+++ scripts/macros/variantRules/talentCreationHandler.js
@@ -1,4 +1,5 @@
-import { actorEditorHelpers } from "../../../utility/actorEditorHelpers.js";
+import { helpers } from "../../helpers.js";
+import { actorEditorHelpers } from "../../utility/actorEditorHelpers.js";
 
 export async function talentItemAddition(item) {
     let ownerActor = item.parent;
@@ -17,6 +18,11 @@ export async function talentItemAddition(item) {
             break;
         case 'Martial Casting':
             await ownerActor.setFlag('midi-qol', 'advantage.concentration', true);
+            break;
+        case 'Tactician':
+            let int = ownerActor.system.abilities.int.mod, wis = ownerActor.system.abilities.wis.mod;
+            let selection = await helpers.dialog('Select an Initative Score', [['Intelligence', 'int'], ['Wisdom', 'wis']], `Your scores are as follows:\n<b>Intelligence:</b> ${int}\n<b>Wisdom:</b> ${wis}\n\nWhich would you like to base your initiative off of?`);
+            await ownerActor.update({'system.attributes.init.ability': selection});
             break;
         default:
             break;